SSCS-10490 moved case access management fields into separate class
