+++ src/main/java/uk/gov/hmcts/reform/sscs/ccd/domain/SscsCaseData.java
@@ -332,6 +332,10 @@ public class SscsCaseData implements CaseData {
 
     @JsonUnwrapped
     @Getter(AccessLevel.NONE)
+    private CaseAccessManagementFields caseAccessManagementFields;
+
+    @JsonUnwrapped
+    @Getter(AccessLevel.NONE)
     private WorkAllocationFields workAllocationFields;
 
     @JsonUnwrapped
@@ -575,6 +579,14 @@ public class SscsCaseData implements CaseData {
             this.postponementRequest = new PostponementRequest();
         }
         return postponementRequest;
+    }
+
+    @JsonIgnore
+    public CaseAccessManagementFields getCaseAccessManagementFields() {
+        if (caseAccessManagementFields == null) {
+            this.caseAccessManagementFields = new CaseAccessManagementFields();
+        }
+        return caseAccessManagementFields;
     }
 
     @JsonIgnore
